+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+# 1.0.0 (2017-07-03)
+
+* Added support to focus row/item index (see [#7](https://github.com/dinony/od-virtualscroll/issues/7))
+* Added readme note on how to use focus row/item feature
+* Added readme note on how to use multiple instances ([#8](https://github.com/dinony/od-virtualscroll/issues/8))
+
+## Breaking Changes
+
+* Removed component input `vsScrollTop`
+* Added single component input `vsUserCmd` to define an observable sequence of `IUserCmd` commands
+
 # 0.2.0 (2017-04-22)
 
 * Initial public release